--- src/containers/new-post/index.js
@@ -1,3 +1,0 @@
-import Post from "./new-post.connector";
-
-export default Post;
