--- src/containers/new-post/new-post.connector.js
@@ -1,8 +1,0 @@
-import { connect } from "react-redux";
-import Post from "./new-post.component";
-
-import { fetchPosts } from "store/thunks/posts";
-
-const mapDispatchToProps = (dispatch) => ({});
-
-export default connect(null, null)(Post);
