+++ src/containers/post/post.component.jsx
@@ -1,86 +1,83 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 import React, { useState, useEffect } from "react";
+
 import { voteForOption, addVoteToUser } from "database/votePost";
-import { fetchPostUser, fetchPostUserAvatar } from "database/postUser";
-import { useSelector } from "react-redux";
+import { fetchPostUserAvatar } from "database/postUser";
 
-const Post = (props) => {
+import PostHeader from "components/post-header";
+import PostFooter from "components/post-footer";
+import PostBody from "components/post-body";
+
+const Post = ({ post, user }) => {
+  console.log("post", post);
+  console.log("user", user);
+  const { userId } = user;
   const {
-    optionA,
-    optionB,
     body,
-    ownerID,
-    hasUserVotedForA,
-    hasUserVotedForB,
-    id,
-    optionAName,
-    optionBName,
-  } = props;
-  const [canUserViewVote, setCanUserViewVote] = useState(false);
-  const [voteACount, setVoteACount] = useState(optionA.length);
-  const [voteBCount, setVoteBCount] = useState(optionB.length);
+    id: postId,
+    owner_username: ownerUsername,
+    total_votes: totalVotes,
+    option_a_name: optionAName,
+    option_b_name: optionBName,
+  } = post;
 
-  const userIDFromState = useSelector((state) => state.users.user.userId);
-  const voteAPercent = (voteACount / (voteACount + voteBCount)) * 100;
-  const voteBPercent = (voteBCount / (voteACount + voteBCount)) * 100;
+  const hasUserVotedForA = post.option_a.includes(user.userId);
+  const hasUserVotedForB = post.option_b.includes(user.userId);
+  const hasUserVoted = hasUserVotedForA || hasUserVotedForB;
 
-  const handleVote = ({ optionName, postId, userId }) => {
-    setCanUserViewVote(true);
-    voteForOption({ optionName, postId, userId });
-    addVoteToUser({ postId, userId });
-    if (optionName === "option_a") {
-      setVoteACount(voteACount + 1);
-    } else {
-      setVoteBCount(voteBCount + 1);
+  const [postOwnerAvatar, setpostOwnerAvatar] = useState("");
+
+  const handleVote = async ({ optionName, postId, userId }) => {
+    try {
+      await voteForOption({ optionName, postId, userId });
+      await addVoteToUser({ postId, userId });
+    } catch (error) {
+      console.error(error);
     }
   };
 
+  // Fetch the avatar of the user who created the post
   useEffect(() => {
-    if (hasUserVotedForA || hasUserVotedForB) {
-      setCanUserViewVote(true);
-    }
-    setVoteACount(optionA.length);
-    setVoteBCount(optionB.length);
-  }, [hasUserVotedForA, hasUserVotedForB]);
+    const fetchAvatar = async () => {
+      const postOwnerAvatar = await fetchPostUserAvatar(ownerUsername);
+      setpostOwnerAvatar(postOwnerAvatar);
+    };
+
+    fetchAvatar();
+  }, []);
 
   return (
-    <li className="rounded-lg pt-1 m-1 mt-3 bg-bluey font-noto tracking-wide">
-      {/* Post Body  */}
-      <h6 className="text-grayy p-2 mx-2 mb-2 rounded-lg border border-grayy font-thin">
-        {body}
-      </h6>
-      <div className="flex justify-center text-bluey ">
-        <button
-          className={`${
-            canUserViewVote ? "hidden" : "show"
-          } bg-pinky w-20 h-10 rounded-l-lg border-r-2 border-gray-600 mb-2 w-1/3`}
-          onClick={() =>
-            handleVote({
-              optionName: "option_a",
-              postId: id,
-              userId: userIDFromState,
-            })
-          }
-        >
-          {optionAName}
-        </button>
-        <button
-          className={`${
-            canUserViewVote ? "hidden" : "show"
-          } bg-orangy w-20 h-10 rounded-r-lg mb-2 w-1/3`}
-          onClick={() =>
-            handleVote({
-              optionName: "option_b",
-              postId: id,
-              userId: userIDFromState,
-            })
-          }
-        >
-          {optionBName}
-        </button>
-      </div>
-      {/* Post Footer  */}
-    </li>
+    <div className="rounded-lg pt-1 m-1 mt-3 bg-bluey font-noto tracking-wide">
+      <PostHeader
+        userData={{
+          username: ownerUsername,
+          postOwnerAvatar,
+        }}
+        voteData={{
+          numOfVotesForOptionA: post.option_a.length,
+          numOfVotesForOptionB: post.option_b.length,
+        }}
+        showVoteResults={hasUserVoted}
+      />
+      <PostBody
+        userData={userId}
+        postData={{ body, postId, optionAName, optionBName }}
+        showVoteResults={hasUserVoted}
+        handleVote={handleVote}
+      />
+      <PostFooter
+        voteData={{
+          hasUserVotedForA,
+          hasUserVotedForB,
+          numOfVotesForOptionA: post.option_a.length,
+          numOfVotesForOptionB: post.option_b.length,
+          totalVotes,
+          optionAName,
+          optionBName,
+        }}
+        showVoteResults={hasUserVoted}
+      />
+    </div>
   );
 };
