+++ src/containers/post/post.connector.js
@@ -1,0 +1,6 @@
+import { connect } from "react-redux";
+import Post from "./post.component";
+
+const mapDispatchToProps = (dispatch) => ({});
+
+export default connect(null, null)(Post);
